fix: make esm mjs
Tools such as `create-react-app` require special configuration to handle
`.cjs` as an extension.  Setting `type: "module"` in `package.json` means
we have to use the `.cjs` extension for commons js which then breaks these
sorts of tools.

This change removes `type: "module"` and uses the `.js` extension for cjs
files and `.mjs` for esm for maximum compatibiliy.